Moved getPath into document itself
