+++ app/src/main/java/net/gsantner/markor/activity/DocumentShareIntoFragment.java
@@ -39,7 +39,6 @@ import net.gsantner.opoc.format.plaintext.PlainTextStuff;
 import net.gsantner.opoc.preference.GsPreferenceFragmentCompat;
 import net.gsantner.opoc.ui.FilesystemViewerAdapter;
 import net.gsantner.opoc.ui.FilesystemViewerData;
-import net.gsantner.opoc.util.StringUtils;
 
 import java.io.File;
 
@@ -276,7 +275,7 @@ public class DocumentShareIntoFragment extends GsFragmentBase {
             if (getActivity() instanceof DocumentActivity) {
                 DocumentActivity a = (DocumentActivity) getActivity();
                 a.setDocument(document);
-                a.showTextEditor(document, null, false, _appSettings.getDocumentPreviewState(StringUtils.getPath(document)));
+                a.showTextEditor(document, null, false, _appSettings.getDocumentPreviewState(Document.getPath(document)));
             }
         }
